requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
requests
requests-oauthlib
google-auth
google-auth-oauthlib
paho-mqtt
celery[redis]
redis
python-dotenv

gunicorn
google-api-python-client
